feat: Add Laser class with length conversion methods
This commit adds a new `Laser` class to the codebase. The class includes methods for converting the length of a laser from centimeters to meters and vice versa. The class also includes a `__str__` method for displaying the length of the laser. This addition provides a convenient way to work with laser lengths in different units of measurement.